(refactor) Moved database models into their own files.
