+++ database-mongo/index.js
@@ -1,6 +1,6 @@
 var mongoose = require('mongoose');
-var MongoClient = require('mongodb').MongoClient;
-var assert = require('assert');
+// var MongoClient = require('mongodb').MongoClient;
+// var assert = require('assert');
 
 // var url = 'mongodb://' + process.env.MONGO_USER + ':' + process.env.MONGO_PASS + '@ds139480.mlab.com:39480/uva'
 var url = 'mongodb://localhost/uva';
@@ -13,66 +13,9 @@ db.on('error', console.error.bind(console, 'connection errorINSIDE BIND'));
 
 db.once('open', function() {
   console.log('connected');
-
-});
-var userSchema = mongoose.Schema({
-  name: String,
-  isOwner: Boolean,
-  isAdmin: Boolean,
-  joined: Date,
-  password: String,
-  salt: String,
-  meta: {
-    reviews: Number,
-    friends: Number
-  }
 });
 
-var reviewSchema = mongoose.Schema({
-  content: String,
-  product: String,
-  rating: Number,
-  username: String,
-  product_id: String,
-  time: Date
-});
-
-var productSchema = mongoose.Schema({
-  name: {type: String, unique: true},
-  year: String,
-  type: String, //e.x. cabernet, merlot
-  redORwhite: String, //red wines or white wines
-  origin: String, //e.x Sonoma County
-  region: String, //e.x California, France
-  priceMin: Number,
-  priceMax: Number,
-  rating: Number, //uva community's average rating
-  apiRating: Number,
-  owner_id: Number
-});
-
-var models = {
-  User: mongoose.model('User', userSchema),
-  Review: mongoose.model('Review', reviewSchema),
-  Product: mongoose.model('Product', productSchema)
-}
-
-module.exports = models;
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+module.exports = db;
 
   //TESTING FUNCTIONALITY
