+++ package.json
@@ -27,11 +27,7 @@
     "express": "^4.15.0",
     "fs": "0.0.1-security",
     "heroku": "^0.2.0",
-    "http": "0.0.0",
-    "https": "^1.0.0",
     "jquery": "^3.1.1",
-    "mongod": "^2.0.0",
-    "mongodb": "^2.2.24",
     "mongoose": "^4.9.0",
     "passport-facebook": "^2.1.1",
     "react": "^15.4.2",
@@ -45,7 +41,6 @@
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "jest": "^19.0.2",
     "mocha": "^3.2.0",
-    "nodemon": "^1.11.0",
     "supertest": "^3.0.0"
   },
   "babel": {
