+++ server/utilities/apiKey.js
@@ -1,0 +1,1 @@
+module.exports = 'd95d33ffb95bc2976aa01b2a520e70ed';
